move make_fd from common to client/helper
